Spheres and Grassmanians bug fixed. Variance parameter added in compact case.

+++ src/spectral_measure.py
@@ -19,9 +19,9 @@ class AbstractSpectralMeasure(torch.nn.Module):
 class MaternSpectralMeasure(AbstractSpectralMeasure):
     def __init__(self, dim, lengthscale, nu, variance=1.0):
         super(MaternSpectralMeasure, self).__init__(dim)
-        self.lengthscale = torch.tensor([lengthscale], device=device, requires_grad=True)
-        self.nu = torch.tensor([nu], device=device)
-        self.variance = Parameter(torch.tensor([variance], device=device, requires_grad=True))
+        self.lengthscale = torch.tensor([lengthscale], device=device, dtype=dtype, requires_grad=True)
+        self.nu = torch.tensor([nu], device=device, dtype=dtype)
+        self.variance = Parameter(torch.tensor([variance], device=device, dtype=dtype, requires_grad=True))
 
     def forward(self, eigenvalue):
         return torch.pow(self.nu[0]/(self.lengthscale[0] ** 2) + eigenvalue, -self.nu[0] - self.dim/4)
@@ -30,7 +30,8 @@ class MaternSpectralMeasure(AbstractSpectralMeasure):
 class SqExpSpectralMeasure(AbstractSpectralMeasure):
     def __init__(self, dim, lengthscale, variance=1.0):
         super(SqExpSpectralMeasure, self).__init__(dim)
-        self.lengthscale = Parameter(torch.tensor([lengthscale], device=device, requires_grad=True))
-        self.variance = Parameter(torch.tensor([variance], device=device, requires_grad=True))
+        self.lengthscale = Parameter(torch.tensor([lengthscale], device=device, dtype=dtype, requires_grad=True))
+        self.variance = Parameter(torch.tensor([variance], device=device, dtype=dtype, requires_grad=True))
+
     def forward(self, eigenvalues):
         return torch.exp((-self.lengthscale[0] ** 2)/2 * eigenvalues)
